requirements.txt
mysql-connector-python
pandas
openpyxl
python-telegram-bot
